Added ACL capabilities to the component and through Joomla, new user group 'teachers'
Fixed editing tests when they were being edited by someone else.
Added alias and created, created_by columns in the db to be auto filled in when test is created.

Fixes issue #7

+++ www/administrator/components/com_tests/views/sessions/view.html.php
@@ -41,10 +41,23 @@ class TestsViewSessions extends JView
 	 */
 	protected function addToolbar()
 	{
-		JToolBarHelper::title('Tests Sessions');
+		$actions = TestsHelper::getActions();
 
-		JToolBarHelper::publish( 'sessions.activate', 'COM_TESTS_ACTIVATE', true );
-		JToolBarHelper::unpublish( 'sessions.deactivate', 'COM_TESTS_DEACTIVATE', true );
-		JToolBarHelper::deleteList( '', 'sessions.delete' );
+		JToolBarHelper::title( JText::_( 'COM_TESTS_VIEW_SESSIONS' ) );
+
+		if ( $actions->get( 'core.edit.state' ) ) {
+			JToolBarHelper::publish( 'sessions.activate', 'COM_TESTS_ACTIVATE', true );
+			JToolBarHelper::unpublish( 'sessions.deactivate', 'COM_TESTS_DEACTIVATE', true );
+		}
+
+		if ( $actions->get( 'core.delete' ) ) {
+			JToolBarHelper::deleteList( '', 'sessions.delete' );
+			JToolBarHelper::divider();
+		}
+
+		if ( $actions->get( 'core.admin' ) ) {
+			JToolBarHelper::divider();
+			JToolBarHelper::preferences('com_tests');
+		}
 	}
 }
